backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
pillow==10.0.1
python-dotenv==1.0.0
transformers==4.40.0
huggingface-hub==0.22.0
speechrecognition==3.10.0
gtts==2.4.0
opencv-python==4.8.1.78
numpy==1.26.4
tensorflow==2.19.0
scikit-learn==1.3.2
